+++ ChangeLog.md
@@ -2,6 +2,9 @@
 IMPROVEMENTS
 - add support multipart upload and multipart download for COS storage, add COS_ALLOW_MULTIPART_DOWNLOAD and COS_CONCURRENCY to config, fix [1074](https://github.com/Altinity/clickhouse-backup/issues/1074) 
 
+BUG FIXES
+- fix one more corner cases for --restore-database-mapping, fix [1146](https://github.com/Altinity/clickhouse-backup/issues/1146#issuecomment-2886456288)
+
 # v2.6.17
 BUG FIXES
 - respect original database engine during `restore --restore-database-mapping`, fix [1146](https://github.com/Altinity/clickhouse-backup/issues/1146)